Resize Life Weeks widget to 1x2, use 4x4 px cells, show on mobile
- Extract shared LifeWeeksGrid component (life-weeks-grid.tsx) with
  configurable cellSizePx and showYearLabels props
- Widget now uses 4x4 px cells (up from 2x2) and delegates to the
  shared grid
- Widget constraint changed from DIM_1x1 to DIM_1x2 on big screen so
  the taller grid fits without heavy scrolling
- life-weeks.tsx refactored to use the shared component
- life-plan.tsx: compact 4x4 px grid (no year labels) shown on mobile
  in a "Life Weeks" section above Milestones
