+++ src/core/jupiter/core/home/widget.py
@@ -340,7 +340,7 @@ WIDGET_CONSTRAINTS = {
     ),
     WidgetType.LIFE_WEEKS: WidgetTypeConstraints(
         allowed_dimensions={
-            HomeTabTarget.BIG_SCREEN: [WidgetDimension.DIM_1x1],
+            HomeTabTarget.BIG_SCREEN: [WidgetDimension.DIM_1x2],
             HomeTabTarget.SMALL_SCREEN: [WidgetDimension.DIM_1x1],
         },
         only_for_workspace_features=[WorkspaceFeature.LIFE_PLAN],
